requirements.txt
einops~=0.6.0
fairscale~=0.4.6
numpy~=1.23.4
similaritymeasures~=0.5.0
torch~=1.12.0+cu113
torchvision~=0.13.0+cu113

mmcv-full~=1.5.0
pyyaml~=6.0
scipy~=1.9.0

loguru~=0.6.0
matplotlib~=3.5.0
prettytable~=3.3.0
shapely~=1.8.4
tqdm~=4.64.0
